Move authed to middleware
